Add test and fix some issues
