+++ src/Sulu/Bundle/AdminBundle/Resources/js/services/ContentLocaleUpdater/index.js
@@ -1,0 +1,5 @@
+// @flow
+import updateRouterAttributesFromUserStoreContentLocale from './updateRouterAttributesFromUserStoreContentLocale';
+import updateUserStoreContentLocaleFromRouterAttributes from './updateUserStoreContentLocaleFromRouterAttributes';
+
+export {updateRouterAttributesFromUserStoreContentLocale, updateUserStoreContentLocaleFromRouterAttributes};
